- Avoid some expensive SDL_GetTicks() calls - Update to squirrel 2.1 - Another rewrite of the scripting code - A few bugfixes I can't remember - Removed Sound object from scripting, replaced it with a global play_music and   play_sound function
SVN-Revision: 3385

+++ src/sprite/sprite.hpp
@@ -91,7 +91,7 @@ private:
 
   float frame;
   int animation_loops;
-  Uint32 last_ticks;
+  float last_ticks;
 
   SpriteData::Action* action;
 };
